PossessAI and much more change to possess things
Now all AI will have ability to be 'paused' in possess case.
All default AI are possessable, maybe for special case they are not (like bossess or Eye of Acherus)

Why replace AI? To avoid multiple check in all AI to handle possess case.
I also removed the AIM_Initialize() call from take Unit::TakePossessOf,because its not the place for it. Instead we need some Reset() function or such.

i have also sync the code as much as possible with wrath

Thank @Phatcat for his help and feedback on this